Moving the membership tests to a unit test
